Tag Creation
Add logic to create tags
