+++ src/App.tsx
@@ -5,31 +5,7 @@ import { Routes, Route, Navigate } from "react-router-dom";
 import { useLocalStorage } from "./hooks/useLocalStorage";
 import { NewNote } from "./NewNote";
 import { v4 as uniqueId } from "uuid";
-
-export type Note = {
-  id: string;
-} & NoteData;
-
-export type RawNote = {
-  id: string;
-} & RawNoteData;
-
-export type RawNoteData = {
-  title: string;
-  markdown: string;
-  tagIds: string[];
-};
-
-export type NoteData = {
-  title: string;
-  markdown: string;
-  tags: Tag[];
-};
-
-export type Tag = {
-  id: string;
-  label: string;
-};
+import { RawNote, NoteData, Tag } from "./types";
 
 function App() {
   const [notes, setNotes] = useLocalStorage<RawNote[]>("NOTES", []);
@@ -44,6 +20,10 @@ function App() {
     });
   }, [notes, tags]);
 
+  /**
+   * @param tags array of tag
+   * @param data object containing current form data
+   */
   function onCreateNote({ tags, ...data }: NoteData) {
     setNotes((prevNotes) => {
       return [
@@ -53,11 +33,27 @@ function App() {
     });
   }
 
+  /**
+   * @param tag object containing tag data
+   */
+  function addTag(tag: Tag) {
+    setTags((prev) => [...prev, tag]);
+  }
+
   return (
     <Container className="my-4">
       <Routes>
         <Route path="/" element={<h1>Notes Home page</h1>} />
-        <Route path="/new" element={<NewNote onSubmit={onCreateNote} />} />
+        <Route
+          path="/new"
+          element={
+            <NewNote
+              onSubmit={onCreateNote}
+              onAddTag={addTag}
+              availableTags={tags}
+            />
+          }
+        />
         <Route path="/:id">
           <Route index element={<h1>Show note Page</h1>} />
           <Route path="edit" element={<h1>Edit Note Page</h1>} />
