+++ src/NewNote.tsx
@@ -1,15 +1,21 @@
-import { NoteData } from "./App";
+import { NoteData, Tag } from "./types";
 import { Noteform } from "./NoteForm";
 
 type NewNoteProps = {
   onSubmit: (data: NoteData) => void;
+  onAddTag: (tag: Tag) => void;
+  availableTags: Tag[];
 };
 
-export function NewNote({ onSubmit }: NewNoteProps) {
+export function NewNote({ onSubmit, onAddTag, availableTags }: NewNoteProps) {
   return (
     <>
       <h1 className="mb-4">New Note</h1>
-      <Noteform onSubmit={onSubmit} />
+      <Noteform
+        onSubmit={onSubmit}
+        onAddTag={onAddTag}
+        availableTags={availableTags}
+      />
     </>
   );
 }
